requirements.txt
discord.py==2.3.2
redis==5.0.1
python-dotenv==1.0.0
flask==3.0.0
asyncio-mqtt==0.16.1
audioop-lts==0.0.1  # Backport pour Python 3.13